requirements.txt
numpy==1.15.4
tensorboardX==1.9
matplotlib==2.2.2
pandas==0.25.2
efficientnet_pytorch
